feature : maximize minimize video view on double tap
