+++ app/src/main/res/layout/fragment_calling.xml
@@ -19,7 +19,7 @@
 
 -->
 <com.waz.zclient.common.controllers.ThemeControllingFrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-             android:layout_width="match_parent"
+    android:layout_width="match_parent"
              android:layout_height="match_parent"
              xmlns:app="http://schemas.android.com/apk/res-auto">
 
@@ -46,6 +46,11 @@
 
     <FrameLayout
         android:id="@+id/controls_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
+
+    <FrameLayout
+        android:id="@+id/full_screen_video_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
