+++ app/src/main/scala/com/waz/zclient/calling/CallingFragment.scala
@@ -17,156 +17,34 @@
  */
 package com.waz.zclient.calling
 
-import android.content.Context
 import android.os.Bundle
 import android.view.{LayoutInflater, View, ViewGroup}
-import android.widget.{FrameLayout, ImageView, TextView}
+import android.widget.FrameLayout
 import androidx.cardview.widget.CardView
 import androidx.gridlayout.widget.GridLayout
-import com.waz.avs.{VideoPreview, VideoRenderer}
-import com.waz.log.BasicLogging.LogTag.DerivedLogTag
-import com.waz.model.{Picture, UserId}
+import com.waz.model.UserId
 import com.waz.service.call.Avs.VideoState
 import com.waz.service.call.CallInfo.Participant
 import com.waz.threading.Threading
 import com.waz.threading.Threading._
 import com.waz.utils.returning
 import com.waz.zclient.calling.controllers.CallController
-import com.waz.zclient.calling.controllers.CallController.CallParticipantInfo
+import com.waz.zclient.calling.views.{OtherVideoView, SelfVideoView, UserVideoView}
 import com.waz.zclient.common.controllers.{ThemeController, ThemeControllingFrameLayout}
-import com.waz.zclient.glide.BackgroundRequest
 import com.waz.zclient.log.LogUI._
 import com.waz.zclient.security.SecurityPolicyChecker
 import com.waz.zclient.utils.ContextUtils._
-import com.waz.zclient.utils.RichView
-import com.waz.zclient.{FragmentHelper, R, ViewHelper}
+import com.waz.zclient.{BuildConfig, FragmentHelper, R}
 import com.wire.signals.Signal
-
-abstract class UserVideoView(context: Context, val participant: Participant) extends FrameLayout(context, null, 0) with ViewHelper {
-  protected lazy val controller: CallController = inject[CallController]
-
-  inflate(R.layout.video_call_info_view)
-
-  private val pictureId: Signal[Picture] = for {
-    z             <- controller.callingZms
-    Some(picture) <- z.usersStorage.signal(participant.userId).map(_.picture)
-  } yield picture
-
-  protected val audioStatusImageView = findById[ImageView](R.id.audio_status_image_view)
-
-  protected val imageView = returning(findById[ImageView](R.id.image_view)) { view =>
-    pictureId.onUi(BackgroundRequest(_).into(view))
-  }
-
-  protected val pausedText = findById[TextView](R.id.paused_text_view)
-
-  private val participantInfoCardView = findById[CardView](R.id.participant_info_card_view)
-
-  protected val stateMessageText = controller.stateMessageText(participant)
-  stateMessageText.onUi(msg => pausedText.setText(msg.getOrElse("")))
-
-  protected val pausedTextVisible = stateMessageText.map(_.exists(_.nonEmpty))
-  pausedTextVisible.onUi(pausedText.setVisible)
-
-  protected val videoCallInfo = returning(findById[View](R.id.video_call_info)) {
-    _.setBackgroundColor(getColor(R.color.black_16))
-  }
-
-   protected val participantInfo: Signal[Option[CallParticipantInfo]] =
-    for {
-      isGroup <- controller.isGroupCall
-      infos   <- if (isGroup) controller.participantInfos() else Signal.const(Vector.empty)
-    } yield infos.find(_.userId == participant.userId)
-
-  protected val nameTextView = returning(findById[TextView](R.id.name_text_view)) { view =>
-    participantInfo.onUi {
-      case Some(p) if p.isSelf => view.setText(getString(R.string.calling_self, p.displayName))
-      case Some(p)             => view.setText(p.displayName)
-      case _                   =>
-    }
-  }
-
-  Signal.zip(controller.isGroupCall, controller.controlsVisible,
-    controller.otherParticipants.map(_.size)
-  ).map {
-    case (true, false, 0 | 1 | 2) => View.GONE
-    case (true, false, _)     => View.VISIBLE
-    case _                    => View.GONE
-  }.onUi(participantInfoCardView.setVisibility)
-
-  protected def registerHandler(view: View) = {
-    controller.allVideoReceiveStates.map(_.getOrElse(participant, VideoState.Unknown)).onUi {
-      case VideoState.Paused | VideoState.Stopped => view.fadeOut()
-      case _                 => view.fadeIn()
-    }
-    view match {
-      case vr: VideoRenderer =>
-        controller.allVideoReceiveStates.map(_.getOrElse(participant, VideoState.Unknown)).onUi {
-          case VideoState.ScreenShare =>
-            vr.setShouldFill(false)
-            vr.setFillRatio(1.5f)
-          case _ =>
-            vr.setShouldFill(true)
-            vr.setFillRatio(1.0f)
-        }
-      case _ =>
-    }
-  }
-
-  Signal.zip(controller.controlsVisible, shouldShowInfo, controller.isCallIncoming).onUi {
-    case (_, true, true) |
-         (false, true, _) => videoCallInfo.fadeIn()
-    case _                => videoCallInfo.fadeOut()
-  }
-
-  val shouldShowInfo: Signal[Boolean]
-}
-
-class SelfVideoView(context: Context, participant: Participant)
-  extends UserVideoView(context, participant) with DerivedLogTag {
-
-    controller.isMuted.onUi {
-        case true => audioStatusImageView.setImageResource(R.drawable.ic_muted_video_grid)
-        case false => audioStatusImageView.setImageResource(R.drawable.ic_unmuted_video_grid)
-      }
-
-  controller.videoSendState.filter(_ == VideoState.Started).head.foreach { _ =>
-    registerHandler(returning(new VideoPreview(getContext)) { v =>
-      controller.setVideoPreview(Some(v))
-      v.setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT))
-      addView(v, 1)
-    })
-  }(Threading.Ui)
-
-  override lazy val shouldShowInfo = Signal.zip(pausedTextVisible, controller.isMuted).map {
-    case (paused, muted) => paused || muted
-  }
-}
-
-class OtherVideoView(context: Context, participant: Participant) extends UserVideoView(context, participant) {
-
-  participantInfo.onUi {
-    case Some(info) =>
-      if (info.isMuted) audioStatusImageView.setImageResource(R.drawable.ic_muted_video_grid)
-      else audioStatusImageView.setImageResource(R.drawable.ic_unmuted_video_grid)
-    case _ =>
-  }
-
-  override lazy val shouldShowInfo = pausedTextVisible
-  registerHandler(returning(new VideoRenderer(getContext, participant.userId.str, participant.clientId.str, false)) { v =>
-    v.setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT))
-    addView(v, 1)
-  })
-}
 
 class CallingFragment extends FragmentHelper {
 
   private val maxVideoPreviews = 12
-  private lazy val controller       = inject[CallController]
-  private lazy val themeController  = inject[ThemeController]
+  private lazy val controller = inject[CallController]
+  private lazy val themeController = inject[ThemeController]
   private lazy val controlsFragment = ControlsFragment.newInstance
-  private lazy val previewCardView  = view[CardView](R.id.preview_card_view)
-
+  private lazy val previewCardView = view[CardView](R.id.preview_card_view)
+  private lazy val fullScreenVideoContainer = view[FrameLayout](R.id.full_screen_video_layout)
   private var viewMap = Map[Participant, UserVideoView]()
 
   private lazy val videoGrid = returning(view[GridLayout](R.id.video_grid)) { vh =>
@@ -179,16 +57,21 @@ class CallingFragment extends FragmentHelper {
       controller.otherParticipants
     ).onUi { case (vrs, selfParticipant, videoCall, incoming, infos, participants) =>
 
-        def createView(participant: Participant): UserVideoView = returning {
-          if (participant == selfParticipant) new SelfVideoView(getContext, participant)
-          else new OtherVideoView(getContext, participant)
-        } { v =>
-          viewMap = viewMap.updated(participant, v)
+      def createView(participant: Participant): UserVideoView = returning {
+        if (participant == selfParticipant) new SelfVideoView(getContext, participant)
+        else new OtherVideoView(getContext, participant)
+      } { v =>
+        viewMap = viewMap.updated(participant, v)
+        if (BuildConfig.CALLING_VVM_MAXIMIZE_MINIMIZE_VIDEO) {
+          v.onDoubleClick.onUi { _ =>
+            showFullScreenVideo(v, selfParticipant, participant)
+          }
         }
+      }
 
-        def findParticipantNameById(userId: UserId): String = infos.find(_.userId == userId).get.displayName
+      def findParticipantNameById(userId: UserId): String = infos.find(_.userId == userId).get.displayName
 
-        val isVideoBeingSent = !vrs.get(selfParticipant).contains(VideoState.Stopped)
+      val isVideoBeingSent = !vrs.get(selfParticipant).contains(VideoState.Stopped)
 
       vh.foreach { v =>
         val videoUsers = vrs.toSeq.collect {
@@ -230,7 +113,7 @@ class CallingFragment extends FragmentHelper {
             findParticipantNameById(v2.participant.userId).toLowerCase
         }.take(maxVideoPreviews)
 
-        gridViews.zipWithIndex.foreach { case (r, index) =>
+        gridViews.zipWithIndex.foreach { case (userVideoView, index) =>
           val (row, col, span, width) = index match {
             case 0 if gridViews.size == 2 => (0, 0, 2, 0)
             case 0 => (0, 0, 1, 0)
@@ -250,14 +133,14 @@ class CallingFragment extends FragmentHelper {
             case _ => GridLayout.CENTER
           }
 
-          r.setLayoutParams(returning(new GridLayout.LayoutParams()) { params =>
+          userVideoView.setLayoutParams(returning(new GridLayout.LayoutParams()) { params =>
             params.width = width
             params.height = 0
             params.rowSpec = GridLayout.spec(row, 1, GridLayout.FILL, 1f)
-            params.columnSpec = GridLayout.spec(col, span, columnAlignment,1f)
+            params.columnSpec = GridLayout.spec(col, span, columnAlignment, 1f)
           })
 
-          if (r.getParent == null) v.addView(r)
+          if (userVideoView.getParent == null) v.addView(userVideoView)
         }
 
         val viewsToRemove = viewMap.filter {
@@ -306,6 +189,33 @@ class CallingFragment extends FragmentHelper {
     super.onDestroyView()
     viewMap = Map()
   }
+
+  def showFullScreenVideo(userVideoView: UserVideoView, selfParticipant: Participant,
+                          participant: Participant): Unit =
+    fullScreenVideoContainer.foreach { container =>
+
+      val nbParticipants = controller.otherParticipants.map(_.size).currentValue.getOrElse(0)
+
+      if (nbParticipants > 2) {
+        var newVideoView: UserVideoView = null
+        if (participant == selfParticipant) newVideoView = new SelfVideoView(getContext, participant)
+        else newVideoView = new OtherVideoView(getContext, participant)
+
+        newVideoView.onDoubleClick.onUi { _ =>
+          videoGrid.foreach(_.addView(userVideoView))
+          container.removeView(newVideoView)
+        }
+
+        videoGrid.foreach(_.removeView(userVideoView))
+        container.addView(newVideoView)
+
+        controller.allVideoReceiveStates
+          .map(_.getOrElse(participant, VideoState.Unknown)).onUi {
+          case VideoState.Started | VideoState.ScreenShare =>
+          case _ => container.removeView(newVideoView)
+        }
+      }
+    }
 }
 
 object CallingFragment {
