+++ build.gradle
@@ -38,6 +38,7 @@ allprojects {
 }
 
 ext {
+    callingVvmMaximizeMinimizeVideo = true
     //Update kotlinCore flag as well when adding new kotlin features
     kotlinSettings = false
     kotlinCore = kotlinSettings
